kazari_op3_defconfig: Disable TSENS
Currently, we are not using this driver for anything, and it isn't actively
being used by us in other MSM drivers (temp alarm driver, which matters the
most, uses the PMIC for temp readings).

Stack corruption was detected in tsens_tm_irq_thread(), but I could not
debug it, so just disable the driver entirely instead as we don't need it.